Replaces the current Synthesis Engine by a port of TinySoundFont (#261)
* Replaced old synth with TinySoundFont

* Fixed some issues on synthesis due to compiler behavior, added copyright notice and thanks to readme for TinySoundFont and SFZero.

* Consider mix volume correctly on synthesis, ensure track volumes are correctly initialized.

* Added TSF and SFZero info in alphaTab output license header

* Fixed compilation issues

* Fixed tests

* Corrected behavior where initial midi events were re-generated after sequencer finish.

* Always execute stop to ensure seek to start

* Ensure midi event dispatching does not exceed playback range

* Fixed metronome

* Fixed issue on GP7 files that midi settings are not correctly loaded.

+++ Source/AlphaTab.CSharp/Platform/CSharp/AlphaSynthWorkerApiBase.cs
@@ -145,7 +145,7 @@ namespace AlphaTab.Platform.CSharp
             DispatchOnWorkerThread(() => { Player.SetChannelSolo(channel, solo); });
         }
 
-        public void SetChannelVolume(int channel, double volume)
+        public void SetChannelVolume(int channel, float volume)
         {
             DispatchOnWorkerThread(() => { Player.SetChannelVolume(channel, volume); });
         }
